% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedHeatmap.R
\name{annotatedCluster}
\alias{annotatedCluster}
\title{Annotated hierarchical clustering}
\usage{
annotatedCluster(mat, scale = T, center = T, annotation = NULL,
  dend_height = unit(5, "cm"), silent = T, ...)
}
\arguments{
\item{mat}{a matrix, with the samples organized in columns.}

\item{scale}{either a logical value or a numeric vector of length equal to the number of columns of ‘x’. Scale the data before clustering.}

\item{center}{either a logical value or a numeric vector of length equal to the number of columns of ‘x’. Center the data before clustering.}

\item{annotation}{a data frame. Each column will be treated as a simple annotation. The data frame must have column names.}

\item{dend_height}{height of the column cluster, should be a ‘unit’ object.}

\item{...}{functions which define complex annotations or vectors of simple annotation. Values should be named arguments to be passed to 'Heatmap' function from 'ComplexHeatmap' package.}
}
\description{
Annotated hierarchical clustering
}
