% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedHeatmap.R
\name{annotatedHeatmap}
\alias{annotatedHeatmap}
\title{Annotated heatmaps}
\usage{
annotatedHeatmap(matrix, ..., column_annotation = NULL,
  row_annotation = NULL, column_names_gp = gpar(fontsize = 7),
  row_names_gp = column_names_gp, row_dend_side = "right",
  row_names_side = "left", heatmap_legend_param = list(color_bar =
  "continuous", legend_height = unit(3, "cm")))
}
\arguments{
\item{matrix}{a matrix. Either numeric or character. If it is a simple vector, it will be converted to a one-column matrix.}

\item{...}{functions which define complex annotations or vectors of simple annotation. Values should be named arguments to be passed to 'Heatmap' function from 'ComplexHeatmap' package.}

\item{column_annotation, row_annotation}{a data frame. Each column will be treated as a simple annotation. The data frame must have column names.}

\item{column_names_gp}{graphic parameters for drawing text.}

\item{row_names_gp}{graphic parameters for drawing text.}

\item{row_dend_side}{should the row cluster be put on the left or right of the heatmap?}

\item{row_names_side}{should the row names be put on the left or right of the heatmap?}

\item{heatmap_legend_param}{a list contains parameters for the heatmap legend. See \code{\link{color_mapping_legend,ColorMapping-method}} for all available parameters.}
}
\description{
Annotated heatmaps
}
