% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedPCA.R
\name{annotatedPCA}
\alias{annotatedPCA}
\title{Annotated PCA plots.}
\usage{
annotatedPCA(...)
}
\arguments{
\item{groups}{a vector with annotation for the features}

\item{main}{a string to be used as title for the plot}
}
\description{
Function to produce PCA plots annotated for one feature.
}
\details{
'annotatedPCA' produces a plot with the first (up to) 6 components, 2 for each scatterplot, coloured following the annotation provided in 'groups'. In the axis of each scatterplot the annotation provided is plotted as scatterplot or as density plot in relation to the coordinates of the annotated component.
}
