% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialMethylation.R
\name{differentialMethylation}
\alias{differentialMethylation}
\title{Differential methylation analysis.}
\usage{
differentialMethylation(M, targets, array.annot.gr, betas = ilogit2(M),
  interest.vars = NULL, outdir = getwd(), ncores = 1, invisible = T)
}
\arguments{
\item{M}{matrix with M values}

\item{targets}{data frame with sample annotation to be used to run limma}

\item{array.annot.gr}{GRanges object with array annotation}

\item{betas}{matrix with Beta values}

\item{interest.vars}{character vector with variables of interest, excluding the confounding factors}

\item{outdir}{path to save results}

\item{ncores}{integer. Number of cores to use to parallellize the DMR analysis}

\item{invisible}{a boolean to determine if the output of 'missMethyl::gometh' is returned}
}
\value{
limma::eBayes
}
\description{
Analysis of differentially methylated positions (DMPs) and detection of differentially methylated regions (DMRs).
}
\author{
Xavier Pastor \email{x.pastorhostench@dkfz.de}
}
