% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMR_detection.R
\docType{methods}
\name{DMR_detection}
\alias{DMR_detection}
\alias{DMR_detection,data.frame,data.frame-method}
\title{Method DMR_detection.}
\usage{
DMR_detection(M, betas, ...)

\S4method{DMR_detection}{data.frame,data.frame}(M, betas, chr, pos,
  significance = 0.05, ncores = 1)
}
\description{
Method DMR_detection.
}
