% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{genotypeSamples}
\alias{genotypeSamples}
\title{Sample genotyping.}
\usage{
genotypeSamples(rgset, outdir)
}
\arguments{
\item{rgset}{a RGChannelSet(Extended) object from 'minfi'}

\item{outdir}{path to output the plots}
}
\description{
Reports the genotypes that can be inferred using the 'rs' probes.
}
