% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gr2bed}
\alias{gr2bed}
\title{GenomicRanges to BED-like data frame.}
\usage{
gr2bed(gr)
}
\arguments{
\item{gr}{a GenomicRanges object}
}
\value{
A data frame with at least six columns with the following info: chromosome, initial position, final position, name of the feature, score and strand.
}
\description{
Function to convert GenomicRanges objects into BED-like data frames.
}
\details{
The function produces a BED-like data frame where the 4th, 5th and 6th columns correspond to the name, the score and the strand associated to the feature. If the 'gr' object has a column called 'name', the name will be taken from this column, otherwise will be set to '.'. For the score, if a column named 'score' is present in the 'gr' object, the value will be taken from this column, otherwise it will be set to '.'. Any additional columns in the 'gr' object will be added as additional annotation from the 7th column on.
}
