% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{guessArrayType}
\alias{guessArrayType}
\title{Guess array type from list of probes.}
\usage{
guessArrayType(x)
}
\arguments{
\item{x}{a character vector with a list of probes to determine the array}
}
\value{
A string with the array description as reported by 'minfi'.
}
\description{
Guesses the array type from a list of probes.
}
