% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{inspector}
\alias{inspector}
\title{Data inspection plotting functions.}
\usage{
inspector(mat, annot, subset = rownames(mat), outdir = getwd(),
  prefix = NULL)
}
\arguments{
\item{mat}{numeric matrix with measures per feature and sample, with samples in the columns and features in the rows}

\item{annot}{data frame with annotation to be plotted.}

\item{subset}{a vector with feature identifiers to cluster samples on a limited number of features}

\item{outdir}{path to output the plots; if 'NULL' then the PCA is printed in the current open device}

\item{prefix}{prefix of the filename to save the file}
}
\description{
Data inspection plotting functions.
}
