% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedPCA.R
\name{pca_report}
\alias{pca_report}
\title{Produces a report with annotated PCA plots.}
\usage{
pca_report(mat, annot = NULL, outdir = NULL, prefix = NULL)
}
\arguments{
\item{mat}{numeric matrix with measures per feature and sample, with samples in the columns and features in the rows}

\item{annot}{data frame with annotation to be plotted.}

\item{outdir}{path to output the plots; if 'NULL' then the PCA is printed in the current open device}

\item{prefix}{prefix of the filename to save the file}
}
\description{
Produces a report with annotated PCA plots.
}
\details{
For each variable in 'annot' a PCA plot is produced. All the PCA plots are the same, only the representation of the annotation changes.
}
