% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotBetaDensityHeatmap}
\alias{plotBetaDensityHeatmap}
\title{Density heatmap}
\usage{
plotBetaDensityHeatmap(mat, filename = NULL, main = NULL)
}
\arguments{
\item{mat}{numeric matrix with measures per feature and sample, with samples in the columns and features in the rows}

\item{filename}{character. File to save the density plot}

\item{main}{title of the density plot}
}
\description{
Density heatmap
}
\details{
'mat' is a numeric matrix of beta values (or values in general in the range [0,1]), with features in rows and samples in columns.

If no filename is provided, the plot will be forwarded to the current graphical device.
}
