% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotControlProbes}
\alias{plotControlProbes}
\title{Control probes plots.}
\usage{
plotControlProbes(rgset, outdir = getwd())
}
\arguments{
\item{rgset}{a RGChannelSet(Extended) object from 'minfi'}

\item{outdir}{path to output the plots}
}
\value{
None
}
\description{
Produces dotplots of the control probes provided by Illumina.
}
\details{
A series of dotplots that may show issues in different steps of the array production are produced.

'staining.png' reports the efficiency of the staining step in each channel. These controls are independent of the hybridization and extension step.

'extension.png' reports the extension efficiency in a hairpin probe, and are therefore sample independent.

'hybridization.png' reports the overall performance of the Infinium Assay using syntetic targets in three different concentrations instead of amplified DNA.  The three different intensities related to the three concentrations should be clearly identifiable.

'target_removal.png' reports the efficiency of the stripping step after the extension. All target removal controls result in low signal compared to the hybridization controls if targets are removed efficiently.

'bisulfite_conversionI.png' and 'bisulfite_conversionII.png' report the efficiency of bisulfite conversion of non-CpG cytosines in the genomic DNA. For type I probes, PM probes should give high signal, while MM probes should give a signal close to background.

'specificityI.png' and 'specificityII.png' report potential nonspecific primer extension and are designed against non polymorphic T sites. For type I probes, PM probes should give high signal, while MM probes should give a signal close to background.

'non_polymorphic.png' reports the overall performance of the assay by querying a particular base in a nonpolymorphic region of the genome, letting you compare assay performance across different samples.

'negative.png' reports the overall background of the signal. The detectin p-value for each probe is estimated based on the intensities of the negative control probes.

See \url{https://support.illumina.com/content/dam/illumina-support/documents/documentation/chemistry_documentation/infinium_assays/infinium_hd_methylation/beadarray-controls-reporter-user-guide-1000000004009-00.pdf} for more details.
}
