% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotSampleCorrelation}
\alias{plotSampleCorrelation}
\title{Plot samples correlation}
\usage{
plotSampleCorrelation(mat, annot = NULL, filename = NULL, main = NULL)
}
\arguments{
\item{mat}{numeric matrix with measures per feature and sample, with samples in the columns and features in the rows}

\item{annot}{data frame with annotation to be plotted.}

\item{filename}{character. File to save the density plot}

\item{main}{title of the density plot}
}
\description{
Plot samples correlation
}
\details{
Pearson correlation between all the samples is calculated.
}
