% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotSamplesDistribution}
\alias{plotSamplesDistribution}
\title{Per sample density plot}
\usage{
plotSamplesDistribution(mat, groups = NULL, filename = NULL,
  main = "Density")
}
\arguments{
\item{mat}{numeric matrix with measures per feature and sample, with samples in the columns and features in the rows}

\item{groups}{vector with annotation for the samples}

\item{filename}{character. File to save the density plot}

\item{main}{title of the density plot}
}
\description{
Per sample density plot
}
\details{
In 'mat', the features are in the rows and the samples in the columns.

If 'groups' is specified, it has to be of the same length and order as the columns in mat, and the density lines will be coloured according to it.

If no filename is provided, the plot will be forwarded to the current device.
}
