% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotSimplifiedQC}
\alias{plotSimplifiedQC}
\title{Median intensities plot.}
\usage{
plotSimplifiedQC(mset, prefix)
}
\arguments{
\item{mset}{a MethylSet(Extended) object from package 'minfi'}

\item{prefix}{prefix of the filename to save the file}
}
\value{
None
}
\description{
Plots the chipwide medians of the Meth and Unmeth channels
}
