% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialMethylation.R
\name{report_DMRs}
\alias{report_DMRs}
\title{Report DMR analysis.}
\usage{
report_DMRs(dmr.gr, array.annot.gr, betas = NULL, annot = NULL,
  prefix = "DMR")
}
\arguments{
\item{dmr.gr}{a GRanges object produced by 'extractRanges' from the 'DMRcate' package}

\item{array.annot.gr}{a GRanges object with array annotation}

\item{betas}{a matrix of beta values}

\item{annot}{a data frame with sample annotation}

\item{prefix}{character, file prefix to store the results}
}
\value{
None
}
\description{
Report the DMR tables in BED format and plots to inspect the results after DMR detection.
}
\details{
DMRs are reported in BED format, sorted by chromosomal location.

In order to produce compatible BED files, 'array.annot.gr' is sorted

'betas' is a matrix with the methylation level (Beta-value) of the CpGs. If 'betas' is provided, a heatmap is produced using the beta values corresponding to the CpG of the top DMRs, given the 'Stouffer', that sum up to 500 CpGs.

If 'annot' is specified, the sample annotation is included in the heatmap.

'prefix' is the file prefix used to store the DMR results and plots.
}
