% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialMethylation.R
\name{report_enrichment}
\alias{report_enrichment}
\title{Report GO/KEGG analysis.}
\usage{
report_enrichment(all.cpg, sig = 0.05, collection = "GO", prefix = NULL,
  invisible = T)
}
\arguments{
\item{all.cpg}{named vector, with significance or any value to subset significant probes}

\item{sig}{threshold used to subset the probes. Probes below or equal to the threshold are selected}

\item{collection}{'GO' or 'KEGG'}

\item{prefix}{file prefix to write the results. It must point to a valid folder}

\item{invisible}{a boolean to determine if the output of 'missMethyl::gometh' is returned}
}
\value{
A data frame with a row for each GO or KEGG term and the following columns:
  \item{Term}{
      GO term if testing GO pathways
  }
  \item{Ont}{
      ontology that the GO term belongs to if testing GO pathways. "BP" - biological process, "CC" - cellular component, "MF" - molecular function.
  }
  \item{Pathway}{
      the KEGG pathway being tested if testing KEGG terms.
  }
  \item{N}{
      number of genes in the GO or KEGG term
  }
  \item{DE}{
      number of genes that are differentially methylated
  }
  \item{P.DE}{
      p-value for over-representation of the GO or KEGG term term
  }
  \item{FDR}{
      False discovery rate
  }
}
\description{
Report GO/KEGG analysis.
}
\details{
This function is a wrapper for the 'gometh' function from 'missMethyl' package that directly writes the results into a folder specified in 'prefix'.
}
\seealso{
missMethyl::gometh
}
