% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Execution of SKIMA.}
\usage{
run(idat_dir = Sys.getenv("IDAT_DIR"),
  sample.annotation = Sys.getenv("SAMPLE_ANNOTATION"),
  blacklist = Sys.getenv("BLACKLIST"), outdir = Sys.getenv("OUTDIR"),
  runCNV = as.logical(Sys.getenv("RUN_CNV")),
  diffMeth = as.logical(Sys.getenv("RUN_DIFFERENTIAL_METHYLATION")),
  removeSNPs = as.logical(Sys.getenv("REMOVE_SNPS")),
  population = Sys.getenv("POPULATION"),
  usePredictedSex = as.logical(Sys.getenv("USE_PREDICTED_SEX")),
  batch.vars = Sys.getenv("BATCH_VARS"),
  seed = as.integer(Sys.getenv("SEED")),
  ncores = as.integer(Sys.getenv("NCORES")))
}
\arguments{
\item{idat_dir}{character. Path to the directory with the idat files.}

\item{sample.annotation}{character. Path to the sample sheet with all the annotation about the samples. Must have a column named 'Sample_Name' with unique sample identifiers, a column named 'Sentrix_ID' with the array identifier and a column named 'Sentrix_Position' with the position in the array where the sample was hybridized.}

\item{blacklist}{character. Path to a file with probes to be flagged or removed from the analysis.}

\item{outdir}{character. Directory where all the results will be writen.}

\item{runCNV}{boolean. Should the CNV analysis be run?}

\item{diffMeth}{boolean. Should the differential methylation analsysis be run?}

\item{removeSNPs}{boolean. Should polymorphic probes with an allellic frequency higher than 1/nrow(sample.annotation) in the population given in 'population' be removed?}

\item{population}{character. A string defining the population where to compare allellic frequencies. One of: NULL, '', 'AFR', 'AMR', 'ASN', 'EUR', 'SAS', 'EAS'. NULL and '' will compare to global allellic frequencies.}

\item{usePredictedSex}{boolean. Should the predicted sex be included in the analysis, and used as confounding variable?}

\item{batch.vars}{character. String with a list of columns from \code{'sample.annotation'} separated by comas to be used as confounding variables.}

\item{seed}{integer. Seed to avoid randomness and make the analysis reproducible.}

\item{ncores}{integer. Number of cores to use for the multithreaded processes.}
}
\value{
None
}
\description{
Runs the SKIMA workflow.
}
\examples{
run()

}
\author{
Xavier Pastor \email{x.pastorhostench@dkfz.de}
}
