% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialMethylation.R
\name{run_limma}
\alias{run_limma}
\title{Run limma on a dataset}
\usage{
run_limma(mat, targets, ...)
}
\arguments{
\item{mat}{numeric matrix}

\item{targets}{data frame with sample annotation}
}
\value{
limma::eBayes
}
\description{
Generation of the fitted model with limma that can be use for any sort of differential analysis, as well as input for DMRcate
}
\details{
The rows in 'targets' must be ordered like the columns in 'mat'. Before the fitted model is produced, 'targets' is checked for variables with absence of replicates and for colinearity.

All the columns present in the 'targets' file will be included in the linear model.
}
